+++ src/compiler/deserialization/mod.ts
@@ -1,7 +1,6 @@
 export * from './list.ucrx-template.js';
 export * from './map.ucrx-entry.js';
 export * from './map.ucrx-template.js';
-export * from './ucd-enable-primitives.js';
 export * from './ucd-entity-feature.js';
 export * from './ucd-feature.js';
 export * from './ucd-function.js';
@@ -9,3 +8,4 @@ export * from './ucd-lib.js';
 export * from './ucd-support-basic.js';
 export * from './ucd-support-defaults.js';
 export * from './ucd-support-non-finite.js';
+export * from './ucd-support-primitives.js';
