+++ src/compiler/deserialization/ucd-support-basic.ts
@@ -1,7 +1,7 @@
 import { ListUcrxTemplate } from './list.ucrx-template.js';
 import { MapUcrxTemplate } from './map.ucrx-template.js';
-import { ucdSupportPrimitives } from './ucd-enable-primitives.js';
 import { UcdSetup } from './ucd-feature.js';
+import { ucdSupportPrimitives } from './ucd-support-primitives.js';
 import { UnknownUcrxTemplate } from './unknown.ucrx-template.js';
 
 export function ucdSupportBasic(setup: UcdSetup): void {
